+++ workspace/solutions/ecmascript/src/0/0/0/1/-.test.ts
@@ -1,3 +1,0 @@
-import { t } from "@/utils/test";
-
-t(import.meta.dir);