+++ workspace/solutions/ecmascript/src/utils/test.ts
@@ -14,7 +14,7 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 // REF: https://bun.sh/docs/cli/test
-import { describe, expect, test } from "bun:test";
+
 import fs from "node:fs";
 import path from "node:path";
 import process from "node:process";
@@ -42,6 +42,7 @@ export interface IExample<
 
 /**
  * 模块测试
+ * @param test - 测试模块
  * @param dir - 目录
  * @param examples - 测试用例
  */
@@ -49,6 +50,7 @@ export async function t<
     I extends any[] = any[],
     O = any,
 >(
+    test: typeof import("bun:test"),
     dir: string,
     examples?: IExample<I, O>[],
 ) {
@@ -60,13 +62,24 @@ export async function t<
     const cwd = paths_.join("/");
 
     if (!examples) {
-        const json = fs.readFileSync(path.join(
+        const examples_file_path = path.join(
             process.cwd(),
             "./../../packages/examples",
             cwd,
             `${id}.json`,
-        ), "utf-8");
-        examples = JSON.parse(json) as IExample<I, O>[];
+        );
+        try {
+            const json = fs.readFileSync(
+                examples_file_path,
+                "utf-8",
+            );
+            examples = JSON.parse(json) as IExample<I, O>[];
+        }
+        catch (error) {
+            void error;
+            console.error(`Examples load failed: ${examples_file_path}`);
+            return;
+        }
     }
 
     const files = fs.readdirSync(dir);
@@ -76,12 +89,15 @@ export async function t<
         func: (await import(`@/${cwd}/${name}`)).default,
     }))()));
 
-    describe.each(solutions)(id, ({ name, func }) => {
+    test.describe.each(solutions)(id, async ({ name, func }) => {
+        // REF: https://github.com/oven-sh/bun/issues/13090
         // !使用异步文件操作后调用 test 会导致其回调无法执行
-        test.each(examples)(`${name} - ${func.name}`, ({ input, output }) => {
-            expect(
+        // !多次调用 t 函数也会导致 test 回调无法执行
+        // *可以通过直接将 test 模块作为参数传入的方式解决
+        test.test.each(examples)(`${name} - ${func.name}`, ({ input, output }) => {
+            test.expect(
                 func(...input),
-                    `${func.name}(${JSON.stringify(input).slice(1, -1)}) != ${JSON.stringify(output)}`,
+                `${func.name}(${JSON.stringify(input).slice(1, -1)}) != ${JSON.stringify(output)}`,
             ).toEqual(output);
         });
     });